migrate Scss Targets
