+++ src/Mobile.BuildTools.FormsCss/Mobile.BuildTools.FormsCss.targets
@@ -2,7 +2,7 @@
 <Project>
 
   <UsingTask TaskName="Mobile.BuildTools.Tasks.ScssProcessorTask"
-             AssemblyFile="$(_MobileBuildToolsParentBuildTasksDll)"/>
+             AssemblyFile="$([System.IO.Path]::Combine('$(MSBuildThisFileDirectory)', 'Mobile.BuildTools.FormsCss.dll'))"/>
 
   <PropertyGroup>
     <CssGDependsOn>
