+++ src/Mobile.BuildTools.FormsCss/buildTransitive/Mobile.BuildTools.FormsCss.targets
@@ -1,0 +1,3 @@
+<Project>
+  <Import Project="$(MSBuildThisFileDirectory)..\build\Mobile.BuildTools.FormsCss.targets" />
+</Project>
